fix: kg decimales en remitos, edicion de remitos ya creados y stock en double

+++ src/main/java/com/tulumcore/api/controllers/BotProductoDTO.java
@@ -8,7 +8,7 @@ public record BotProductoDTO(
         String nombre,
         String descripcion,
         Double precio,
-        Integer cantidadStock,
+        Double cantidadStock,
         String medidas,
         String imageUrl
 ) {}
